Bugfix for Issue #13
Nil values are not supported by go for basic types. Therfore f lua passes
nil values as arguments to go functions, go now uses the "default values"
for uninitialized variables (e.g. 0 for int, false for bool, "" for string).

Modified handling for argument types which are not currently supported.
In this case a panic is raised, which is recovered by the "call" function, which allows
to return a proper error (via chan).

Bugfix for https://github.com/beatgammit/luna/issues/13